Changed Ident from int to long

+++ src/src/org/tms/api/Subset.java
@@ -22,7 +22,7 @@ import org.tms.api.io.IOOption;
 public interface Subset extends TableElement
 {
     public String getUuid();
-    public int getIdent();  
+    public long getIdent();  
     
     /**
      * Return the number of {@link Column}s in this {@code Subset}.
